requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.34.0
bcrypt==3.2.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.115.12
graphql-core==3.2.6
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
motor==3.7.1
openai==1.84.0
packaging==25.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
strawberry-graphql==0.270.5
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
